CreateInterpreter.c
[R/r] [params]: read a rectangle. params of a rectangle is composed of 4 integers:
[upper left corner x coord, upper left corner y coord, width, height]
[S/s] [params]: read a square. params of a square is composed of 3 integers:
[upper left corner x coord, upper left corner y coord, width]
[P/p] [command]: print the requested data according to the following commands:
-D/d: print all the rectangles and the squares previously read. Print them in this format:
[upper left x coord, upper left y coord, lower right x coord, lower right y coord]
 -R/r [index]: print the area of the rectangle specified with the index parameter
-I/i : print total intersection area of the given rectangle/square set
-N/n : print net surface area of all rectangles/squares.
-B/b : print the bounding box coordinates of the data in this format:
[upper left x coord, upper left y coord, lower right x coord, lower right y coord]
[Q/q] : terminate the program.

----------------------------------------------------------

The number of geometric objects, i.e. rectangles and squares, cannot exceed 100.
 In the Cartesian Coordinates, a region cannot be covered by more than two rectangles at the
same time. In other words, there can be regions where two rectangles overlap, but not three
at the same time.
 Net surface area is the coverage of the planar region occupied by all rectangles/squares.
Therefore, a region that is covered with more than one rectangle is counted just once. (i.e.
net surface area of two rectangles A and B is Area(A) + Area(B) – Intersection (A , B)).
 The interpreter is case insensitive. Hence, all commands can arbitrarily be capitalized (or
not).
 There are three error cases that the interpreter handles and reports (refer to input4.txt and
output4.txt for exemplary error cases):
o When a command is not recognized
o When a print command is wrong
o When a rectangle index is invalid

-----------------------------------------------------------------

Store a set of attributes for rectangles or squares in the
memory and calculate some useful geometric functions using them.