Create SimpleVocabScanner.cpp
Reads a vocabulary, i.e. a set of words, from the std. input and (1) counts the number of occurrence of each word in the vocabulary, (2) prints out the words in the vocabulary in ascending order. The vocabulary size will not exceed 1000.

All words in the vocabulary should be treated case insensitively; and the words must be printed in small letters.