Create Pen.cpp - Pen on a Turtle
A mechanical turtle that walks around the room under the control of a C++ program. The turtle holds a pen in one of two positions, up or down. While the pen is down, the turtle traces out shapes as it moves; while the pen is up, the turtle moves about freely without writing anything. In this problem, you will simulate the operation of the turtle and create a computerized sketchpad as well. Use a 20-by-20 array floor which is initialized to zeros.

Command Meaning
1 Pen up
2 Pen down
3 Turn right
4 Turn left
5 10 Move forward 10 spaces (or a number other than 10)
6 Print the 20-by-20 array
9 End of data