Create ColorDragOpenGL.cpp
A rectangle on the left which is drawn only with its borders in black.
A blue triangle on the right.

When the user clicks on the triangle, its color should change to a different color. And this
should happen again and again as the user continues to click. You are free to develop your
own method to decide the next color.

The user should be able to drag the rectangle around and place it in a new position with the
mouse.