Create SpinningCubeOpenGL.cpp
The cube should rotate after a key from keyboard is pressed. The cube should rotate to right (d
key), left (a key), up (w key) and down (s key) when the corresponding key is pressed.

Some faces should be colored with fixed color while some of them are colored with a
mixture of colors.